Fix #82 - add DM58 testing at Part 12 Step 12 -> review of the document shows the work under step 12 was actually and add on to step 11
