+++ src/org/etools/j1939_84/controllers/part12/Part12Controller.java
@@ -39,8 +39,7 @@ public class Part12Controller extends PartController {
              new Part12Step08Controller(),
              new Part12Step09Controller(),
              new Part12Step10Controller(),
-             new Part12Step11Controller(),
-             new Part12Step12Controller());
+             new Part12Step11Controller());
     }
 
     /**
